rust: split openssl facade shape
Defend the Rust OpenSSL facade shape invariant without changing OpenSSL behavior or public error strings.

Validation: cargo test -p rubin-consensus verify_sig_openssl; cargo clippy -p rubin-consensus; full rubin-consensus tests; OpenSSL isolation checker.
